pyproject.toml
[project]
name = "coralation-analisis"
version = "0.1.0"
description = "Automated bias/shortcut discovery tool for image classification models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.45.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.18.0",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "qwen-vl-utils>=0.0.8",
    "torchcam>=0.4.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
    "tqdm>=4.66.0",
    "einops>=0.7.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "opencv-python-headless>=4.9.0",
    "matplotlib>=3.8.0",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "gpu: GPU integration tests (require --gpu flag to run)",
]

[project.scripts]
coralation = "main:main"
